Moving GlrmLoss into the genmodel package
