+++ h2o-algos/src/main/java/hex/glrm/GLRMModel.java
@@ -1,6 +1,7 @@
 package hex.glrm;
 
 import hex.*;
+import hex.genmodel.algos.glrm.GlrmLoss;
 import hex.genmodel.algos.glrm.GlrmRegularizer;
 import hex.svd.SVDModel.SVDParameters;
 import water.*;
