+++ h2o-algos/src/main/java/hex/schemas/GLRMV3.java
@@ -4,7 +4,7 @@ import hex.DataInfo;
 import hex.genmodel.algos.glrm.GlrmRegularizer;
 import hex.glrm.GLRM;
 import hex.glrm.GLRMModel.GLRMParameters;
-import hex.glrm.GlrmLoss;
+import hex.genmodel.algos.glrm.GlrmLoss;
 import hex.svd.SVDModel.SVDParameters;
 import water.api.API;
 import water.api.schemas3.ModelParametersSchemaV3;
